refactor#15: 설정 코드 병합
